+++ packages/react-zeugma/src/index.ts
@@ -1,5 +1,6 @@
-export { DashboardProvider } from './entities/dashboard'
+export { Zeugma } from './entities/dashboard'
 export type {
+  ZeugmaProps,
   ZeugmaClassNames,
   DashboardStateValue,
   DashboardActionsValue,
@@ -10,5 +11,5 @@ export { PaneTree } from './widgets/pane-tree'
 export { ResizableContainer } from './widgets/resizable-container'
 export type { ResizableContainerProps } from './widgets/resizable-container'
 export { Pane, DragHandle } from './entities/pane'
-export type { PaneRenderProps } from './entities/pane'
+export type { PaneProps, DragHandleProps, PaneRenderProps } from './entities/pane'
 export * from './shared'